Separation functions from FeatureContext.php into proper file
