+++ app/features/bootstrap/FeatureContext.php
@@ -20,41 +20,5 @@ class FeatureContext implements Context
      */
     public function __construct()
     {
-        $this->domainList = [];
-    }
-
-    /**
-     * @Given The email adress provided by the user is :email
-     */
-    public function theEmailAdressProvidedByTheUserIs($email)
-    {
-        $pattern = '/\S+@\S+\.\S+/';
-
-        PHPUnit\Framework\Assert::assertRegExp($pattern, $email);
-    }
-
-    /**
-     * @When The user click the :button button
-     */
-    public function theUserClickTheButton($button)
-    {
-        PHPUnit\Framework\Assert::assertTrue(true, $button == "register");
-    }
-
-    /**
-     * @Then The user account is validated if :email compute to a recognized domain
-     */
-    public function theUserAccountIsValidatedIfComputeToARecognizedDomain($email)
-    {
-        $domain = explode("@", $email);
-        PHPUnit\Framework\Assert::assertContains($domain[1], $this->domainList);
-    }
-
-    /**
-     * @Given :domain is part of the authorized domains
-     */
-    public function isPartOfTheAuthorizedDomains($domain)
-    {
-        array_push($this->domainList, $domain);
     }
 }
